Make the text query apply for every load of data
Previously the text query was only applied when commiting the change to
the text box, but it was ignored when other actions caused a reload. Now,
the text query is applied as part of the normal data loading, so it is
always active. Because of this change, bad queries no longer trigger an
alert, but only change the query's text color to red.

Signed-off-by: Jonathan Mendez <br85ur3pqb@snkmail.com>