Make tag filter apply before paging happens
The tag filter was previously implemented as a filter function that was
applied on the outer treegrid store. This meant that it was not applied
during the inner store's load, so paging was messed up. Now the tag filter
is implemented as a QueryFilter, which does get properly set on the inner
store and is applied during the load.

Signed-off-by: Jonathan Mendez <br85ur3pqb@snkmail.com>